Collections in the response are no longer Options
This library comes from the integration tests of the content api.
In those tests, it was important to distinguish between an
element that was completely missing, and one that just had
no children.

This isn't true for most end users of this library and
introduces needless Option checks, so we no longer distinguish
between the two cases.