Support cross compile for 2.8.0 and 2.8.1. Enable snapshot releases.
See http://code.google.com/p/simple-build-tool/wiki/CrossBuild for
information on cross building in sbt.  This project defaults to
just building for 2.8.1 now, prefix any command with + to build
for 2.8.0 and 2.8.1.

So you'll probably want to do +update and +publish (or +publish-local)
from now on to publish.

I've also added a hack to the project file to enable publishing
snapshot releases to the guardian repo, which should help avoid
uncessary verion number inflation.