Add Edition to ItemResponse class
Section item endpoints (e.g. /business) not return an Edition object.
This has been added to the ItemResponse class.

Note, it is not possible to access this property yet. This is because
an 'edition' method is mixed in to the ItemQuery class as part of the
ShowParameters trait. The same problem already exists for section and
tag. This is a general problem with the design of the Api Client.