Define and generate the Content API model via Scrooge using Thrift.
This substantial change defines the Content API model in a thrift file.
This allows us to have the client facing CAPI model defined in one place, and one place only.

This has a number of benefits, namely:
 - Only one declaration to maintain for developers on the CAPI team.
 - Allows us to formerly declare our model (with comments) to clients.
 - Allows us to remove nasty Map[String, String], replacing with Thrift structs that allow us to declare types.

In order to achieve this, improvements had to be made to Json4s, namely:
 - PR: https://github.com/json4s/json4s/pull/293
 - PR: https://github.com/json4s/json4s/pull/288

These improvements allow for deserialization directly to Scrooge generated code as Scrooge does not generate case classes meaning json4s could not deserialize. Hopefully this should
prove of use to many!

In order facilitate the changes to Json4s, we have to roll with our own unmanaged lib inside to content-api-scala-client project for now.
Once a new version is released with our changes, we can move back to the former.