Add `Format` to the Scala CAPI client
Rendering platforms have moved to use the `Format` type to describe the
format that rendered data should take. This includes three specific
fields:
 - `Display`
 - `Design`
 - `Theme (Pillar/Special)`

Currently the platforms handle the decision logic around how to
construct this object on platform but ideally this will be done upstream
in order to standardise across platforms and to simplify the parsing of
data.

In this commit I upstream the `Format` type to the CAPI Scala client.

Changes:
- Introduce the `Display`, `Theme` and `Design` traits and their
associated `case object`s.
- Refactor `CAPIModelEnrichment` to move code shared by `RichContent`
and `Format` to the upper scope.
- Reproduce the decision logic used by platforms to construct the
`Format` object.