Remove superfluous/confusing `version.sbt` files
The current releases of `content-api-client` on Maven Central are
v18.0.2 :

https://repo1.maven.org/maven2/com/gu/content-api-client_2.13/18.0.2/

...and we have a top-level `version.sbt` file that agrees (pointing
at the next snapshot version, `18.0.3-SNAPSHOT`):

https://github.com/guardian/content-api-scala-client/blob/main/version.sbt#L1

However, there are also these `version.sbt` files in subfolders that
override that version to `17.24.0-SNAPSHOT` ??:

https://github.com/guardian/content-api-scala-client/blob/1a123d96af9dc2476c69589d2859f373d967dcaa/client-default/version.sbt#L1
https://github.com/guardian/content-api-scala-client/blob/1a123d96af9dc2476c69589d2859f373d967dcaa/client/version.sbt#L1

If you check out `main` on this repo at the moment and run
`sbt ++publishLocal`, the version of artifacts that are published will
be `17.24.0-SNAPSHOT`, which is not helpful :( If you are trying to
test out your locally-built client with other projects that use the
CAPI client, the other projects may end up selecting published versions
of the CAPI client with more-recent version numbers, rather than the
locally-built version you want to test.

I can't see why these sub-versions need to exist, and think it would be
better if they were removed so there's no ambiguity to the version?

I realise that there is version differential with https://github.com/guardian/content-api-models,
which is fine, it's a different repo and I think I understand that, but
not why these old version numbers are hanging around in
https://github.com/guardian/content-api-scala-client ?