To apply latest version of capiModelsVersion.
 This is to update content-api-scala-client to avoid error between latest and old version of Capi Models.

Related to the work of Scala Steward integration in apple-news code. We are encountering error NoSuchMethodError at the moment in AN code after upgrading latest versions of both content-scala-client and content-api-models which suggests that scala-client is not updated with latest models hence applying the change to release updated version of scala-client.
(Guided by Roberto Tyley)